pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.8",
    "langchain[anthropic]>=0.3.26",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
]
